Updated tick parameters to double for EmStd

+++ table-api/src/main/java/io/deephaven/api/updateby/UpdateByOperation.java
@@ -625,7 +625,7 @@ public interface UpdateByOperation {
      * @param pairs The input/output column name pairs
      * @return The aggregation
      */
-    static UpdateByOperation EmStd(long tickDecay, String... pairs) {
+    static UpdateByOperation EmStd(double tickDecay, String... pairs) {
         return EmStdSpec.ofTicks(tickDecay).clause(pairs);
     }
 
@@ -649,7 +649,7 @@ public interface UpdateByOperation {
      * @param pairs The input/output column name pairs
      * @return The aggregation
      */
-    static UpdateByOperation EmStd(final OperationControl control, long tickDecay, String... pairs) {
+    static UpdateByOperation EmStd(final OperationControl control, double tickDecay, String... pairs) {
         return EmStdSpec.ofTicks(control, tickDecay).clause(pairs);
     }
